requirements.txt
pytorch-lightning
hydra-core
av2
ray
tensorboard